Fixed Heapify.py
Heapify the children first and then swap the data -> then heapify to maintain the heap property. If we don't heapify the children first, then if the original root is bigger than its 2 children, it'll never swap

EX.
4 
 - 2
     - 1
     - 10
 - 3
     - 7

Old heapify would result in 
4 
 - 2
     - 1
 - 10
     - 7
     - 3

because largest_child would choose between 2 and 3 since the subtrees weren't heapified, but since we heapify the subtrees first, largest_child is between 2 and 10 and then can swap properly and then heapify the subtree which swapped.